+++ source/extensions/omni.isaac.ros2_bridge-humble/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.12.0] - 2023-02-28
+### Fixed
+- Issue where image based messages always used increasing simulation time
+### Added
+- resetSimulationTimeOnStop to Camera Helper node. 
 
 ## [1.11.2] - 2023-02-21
 ### Fixed